Intermediate commit, serial communication was set up, config files hosted up, implemented new version of WeatherDataPool, created sketch for testing some features

+++ src/libraries/SerialMessageTransfer/SerialMessageTransfer.h
@@ -7,7 +7,7 @@ class SerialMessageTransfer
 {
 public:
     static String WrapMessage(const String &message);
-    static uint8_t GetNewMessagesCount();
+    static size_t GetNewMessagesCount();
     static bool CheckNewMessages(const String &input);
     static String GetLastMessage(bool removeFromStack = true);
     static void ClearMessages();
